+++ src/main/java/firok/tiths/common/Entities.java
@@ -2,6 +2,7 @@ package firok.tiths.common;
 
 import firok.tiths.entity.projectile.ProjectileDashingStar;
 import firok.tiths.entity.special.EnderBeacon;
+import firok.tiths.entity.trans.Alchemy;
 import firok.tiths.util.reg.RegEntity;
 
 @SuppressWarnings("all")
@@ -16,6 +17,9 @@ public class Entities
 //
 //	@RegEntity(network=301)
 //	public static TransformingCloud transforming_cloud;
+	@RegEntity(network = 302)
+	public static Alchemy alchemy;
+
 	@RegEntity(network = 350)
 	public static EnderBeacon ender_beacon;
 }