+++ src/main/java/firok/tiths/util/Predicates.java
@@ -6,8 +6,6 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
 import java.util.Random;
@@ -71,7 +69,7 @@ public final class Predicates
 	}
 	public static boolean canTrigger(Random rand,double rate)
 	{
-		return rand.nextDouble()<rate;
+		return rate > 0 && rand.nextDouble() < rate;
 	}
 
 	public static boolean canOreGenWorld(IBlockState state)